First part completed
The game board has 8x8 squares (initially, all represented by question marks)
There are 8 diamonds hidden on the board, each diamond behind one of the squares
When the user clicks on a square
If the square was hiding a diamond, the diamond appears
Otherwise, the square is opened, and blank
The game ends when all diamonds are found. The user's score is the number of squares still left unturned.